fix: apply stricter mypy type casts for CI compatibility
- Use np.array() constructor instead of np.asarray() for stricter stubs
- Add explicit float() casts for np.mean returns in heat_capacity.py
- Use list() conversion for np.percentile in freeman_carroll.py
- Use broad 'type: ignore' for pywt import to handle both installed/uninstalled
- Fix Figure type assignment in model_specific_plots.py

+++ src/pkynetics/technique_analysis/dsc/heat_capacity.py
@@ -297,10 +297,10 @@ class CpCalculator:
                 region_slice = slice(start_idx, end_idx)
 
                 # Calculate regional averages
-                temp = np.mean(temperature[region_slice])
-                sample_signal_avg = np.mean(heat_flow[region_slice])
-                ref_signal_avg = np.mean(ref_heat_flow[region_slice])
-                ref_cp_value = np.mean(ref_cp[region_slice])
+                temp = float(np.mean(temperature[region_slice]))
+                sample_signal_avg = float(np.mean(heat_flow[region_slice]))
+                ref_signal_avg = float(np.mean(ref_heat_flow[region_slice]))
+                ref_cp_value = float(np.mean(ref_cp[region_slice]))
 
                 # Calculate Cp
                 cp = self._calculate_regional_cp(
